// src/types.ts
export type InfoType = {
    name: string;
    email: string;
    phone: string;
    intro: string;
    firstHeader: string;
    firstSubHeader: string;
    firstTextOne: string;
    firstTextTwo: string;
    secondHeader: string;
    secondSubHeader: string;
    secondTextOne: string;
    secondTextTwo: string;
  };